requirements.txt
fastapi==0.115.8
SQLAlchemy==2.0.38
alembic==1.14.1
asyncpg==0.30.0
uvicorn==0.34.0
pydantic==2.10.6
pydantic[email]
environs==14.1.0
redis==5.2.0
pyjwt==2.10.1
passlib==1.7.4
bcrypt==4.0.1
sqladmin==0.20.1
itsdangerous==2.1.2
boto3==1.37.7
fastapi-cache2==0.2.2